Adding some features to admin-app module

+++ backend/src/domain/admin/admin.repository.ts
@@ -17,35 +17,22 @@ export class AdminRepository extends Repository<Admin> {
   }
 
   async findAllAdmins(): Promise<IAdmin[]> {
-    return this.find();
+    return this.find({ relations: ['user'] });
   }
 
   async findAdminById(id: number): Promise<IAdmin | null> {
-    return this.findOne({ where: { user_id: id } });
+    return this.findOne({ where: { user_id: id }, relations: ['user'] });
   }
 
   async findAdminByEmail(email: string): Promise<IAdmin | null> {
-    return this.createQueryBuilder('admin')
-      .innerJoinAndSelect('admin.user', 'user', 'user.email = :email', { email })
-      .getOne();
+    return this.createQueryBuilder('admin').innerJoinAndSelect('admin.user', 'user', 'user.email = :email', { email }).getOne();
   }
 
   async updateAdmin(user_id: number, updateAdminDto: UpdateAdminDto): Promise<UpdateResult> {
-    return this.createQueryBuilder()
-      .update(Admin)
-      .set(updateAdminDto)
-      .where('user_id = :user_id', { user_id })
-      .execute();
+    return this.createQueryBuilder().update(Admin).set(updateAdminDto).where('user_id = :user_id', { user_id }).execute();
   }
 
   async deleteAdmin(user_id: number): Promise<DeleteResult> {
-    return this.manager
-      .getRepository(User)
-      .createQueryBuilder()
-      .delete()
-      .from(User)
-      .where('id = :id', { id: user_id })
-      .returning('*')
-      .execute();
+    return this.manager.getRepository(User).createQueryBuilder().delete().from(User).where('id = :id', { id: user_id }).returning('*').execute();
   }
 }
